refactor(admin): co-locate nav label/icon helpers with adminNavigation
References PR #119 review comments r3107535868 and r3107535906 from @Jorenve.
